requirements.txt
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
automium==0.2.6
automium-web==0.1.1
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
crispy-bootstrap4==2024.1
decorator==5.1.1
defusedxml==0.7.1
Django==5.0.7
django-crispy-forms==2.2
django-debug-toolbar==4.4.6
django-filter==24.2
djangorestframework==3.14.0
docopt==0.6.2
docutils==0.21.2
drf-spectacular==0.27.2
executing==2.0.1
fastjsonschema==2.19.1
flake8==7.0.0
flake8-docstrings==1.7.0
Flask==3.0.3
gunicorn==23.0.0
idna==3.7
inflection==0.5.1
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
legacy==0.1.7
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mccabe==0.7.0
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
packaging==24.0
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
platformdirs==4.2.2
prompt_toolkit==3.0.45
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.1
pydocstyle==6.3.0
pyflakes==3.2.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
requests==2.32.3
rpds-py==0.18.1
sentry-sdk==2.12.0
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.5
sqlparse==0.4.2
stack-data==0.6.3
tinycss2==1.3.0
tornado==6.4
traitlets==5.14.3
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.3
yarg==0.1.9
